+++ paddle/fluid/pybind/eager.cc
@@ -21,14 +21,13 @@ limitations under the License. */
 #include "paddle/fluid/memory/memcpy.h"
 #include "paddle/fluid/platform/enforce.h"
 #include "paddle/fluid/pybind/eager.h"
-#include "paddle/fluid/pybind/eager_op_function_impl.h"
 #include "paddle/fluid/pybind/eager_utils.h"
-#include "paddle/fluid/pybind/op_function_common.h"
 #include "paddle/pten/api/include/core.h"
 #include "paddle/pten/common/data_type.h"
 #include "paddle/pten/core/convert_utils.h"
 #include "paddle/pten/core/dense_tensor.h"
 
+// #include "paddle/fluid/pybind/eager_op_function_impl.h"
 namespace paddle {
 namespace pybind {
 
@@ -131,7 +130,7 @@ void BindEager(pybind11::module* module) {
   }
 
   BindFunctions(m.ptr());
-  BindEagerOpFunctions(&m);
+  // BindEagerOpFunctions(&m);
 }
 
 }  // namespace pybind
